+++ src/main/java/lagatrix/gui/window/MainWindow.java
@@ -3,6 +3,7 @@ package lagatrix.gui.window;
 import java.awt.Color;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import javax.swing.ImageIcon;
 import lagatrix.connection.RequesterManager;
 import lagatrix.connection.communicators.AESCommunicator;
 import lagatrix.entities.dto.Connection;
@@ -192,6 +193,11 @@ public class MainWindow extends javax.swing.JFrame {
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setTitle("Lagatrix");
+        setIconImage(new ImageIcon(getClass().getResource("/logo.png")).getImage());
+        setMaximumSize(new java.awt.Dimension(1101, 606));
+        setMinimumSize(new java.awt.Dimension(1101, 606));
+        setResizable(false);
+        setSize(new java.awt.Dimension(1101, 606));
 
         mainPanel.setBackground(new java.awt.Color(245, 245, 245));
         mainPanel.setRadius(61);