+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LACLI003 - 11/05/2023 - Implements poweroff and reboot machine.
 - LACLI008 - 12/05/2023 - Change to string method of entity.
 - LACLI008 - 14/05/2023 - Implements get system information.
+- LACLI008 - 15/05/2023 - Add logo in window icon.
 
 ### Fixed
 - LACLI004 - 06/05/2023 - Selected rows bugs.